aula-03 completa

+++ server/src/routes/memories.ts
@@ -3,8 +3,15 @@ import { z } from 'zod'
 import { prisma } from '../lib/prisma'
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get('/memories', async () => {
+  app.addHook('preHandler', async (request) => {
+    await request.jwtVerify()
+  })
+
+  app.get('/memories', async (request) => {
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: request.user.sub,
+      },
       orderBy: {
         createdAt: 'asc',
       },
@@ -14,17 +21,85 @@ export async function memoriesRoutes(app: FastifyInstance) {
       return {
         id: memory.id,
         coverUrl: memory.coverUrl,
-        excerpt: memory.content.substring(0, 115).concat('...'),
+        except: memory.content.substring(0, 115).concat('...'),
       }
     })
   })
 
-  app.get('/memories/:id', async (request) => {
-    const paramsSchema = z.object({
+  app.get('/memories/:id', async (request, reply) => {
+    const paramsShema = z.object({
       id: z.string().uuid(),
     })
+    const { id } = paramsShema.parse(request.params)
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
 
-    const { id } = paramsSchema.parse(request.params)
+    if (!memory.isPublic && memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
+    return memory
+  })
+
+  app.post('/memories', async (request) => {
+    const bodyShema = z.object({
+      content: z.string(),
+      coverUrl: z.string(),
+      isPublic: z.coerce.boolean().default(false),
+    })
+    const { content, isPublic, coverUrl } = bodyShema.parse(request.body)
+    const memory = await prisma.memory.create({
+      data: {
+        content,
+        coverUrl,
+        isPublic,
+        userId: request.user.sub,
+      },
+    })
+    return memory
+  })
+
+  app.put('/memories/:id', async (request, reply) => {
+    const paramsShema = z.object({
+      id: z.string().uuid(),
+    })
+    const { id } = paramsShema.parse(request.params)
+
+    const bodyShema = z.object({
+      content: z.string(),
+      coverUrl: z.string(),
+      isPublic: z.coerce.boolean().default(false),
+    })
+    const { content, isPublic, coverUrl } = bodyShema.parse(request.body)
+
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
+
+    memory = await prisma.memory.update({
+      where: { id },
+      data: {
+        content,
+        coverUrl,
+        isPublic,
+      },
+    })
+    return memory
+  })
+
+  app.delete('/memories/:id', async (request, reply) => {
+    const paramsShema = z.object({
+      id: z.string().uuid(),
+    })
+    const { id } = paramsShema.parse(request.params)
 
     const memory = await prisma.memory.findUniqueOrThrow({
       where: {
@@ -32,65 +107,9 @@ export async function memoriesRoutes(app: FastifyInstance) {
       },
     })
 
-    return memory
-  })
-
-  app.post('/memories', async (request) => {
-    const bodySchema = z.object({
-      content: z.string(),
-      coverUrl: z.string(),
-      isPublic: z.coerce.boolean().default(false),
-    })
-
-    const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
-
-    const memory = await prisma.memory.create({
-      data: {
-        content,
-        coverUrl,
-        isPublic,
-        userId: '72d4827b-4650-4eeb-9a9b-b16e4f85a0b8',
-      },
-    })
-
-    return memory
-  })
-
-  app.put('/memories/:id', async (request) => {
-    const paramsSchema = z.object({
-      id: z.string().uuid(),
-    })
-
-    const { id } = paramsSchema.parse(request.params)
-
-    const bodySchema = z.object({
-      content: z.string(),
-      coverUrl: z.string(),
-      isPublic: z.coerce.boolean().default(false),
-    })
-
-    const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
-
-    const memory = await prisma.memory.update({
-      where: {
-        id,
-      },
-      data: {
-        content,
-        coverUrl,
-        isPublic,
-      },
-    })
-
-    return memory
-  })
-
-  app.delete('/memories/:id', async (request) => {
-    const paramsSchema = z.object({
-      id: z.string().uuid(),
-    })
-
-    const { id } = paramsSchema.parse(request.params)
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
 
     await prisma.memory.delete({
       where: {
